cpio: revert previous commit
I have done a little bit of observation and that change is not as easy as it
could look.

Firstly - asking for %{_sbindir}/rmt is not OK because we have older systems
like RHEL6- which has only /sbin/rmt and not /usr/sbin/rmt.  Secondly, I
looked at some older debian system, and it does not have the /sbin/rmt (but
has only /usr/sbin/rmt).

Looking at star source code, star based utilities always try to look for
/etc/rmt (which, btw seems to exist on all platforms).

~> For that ^^ reasons, I'm switching back to /etc/rmt

This reverts commit 1518d48b252433e1e311be8866239ad0135b19bc

Version: 2.11-20